feature: mlaprolog open



# message auto-generated for no-merge-commit merge:
!2280 merge master into master

feature: mlaprolog open

Created-by: jiyingying
Commit-by: jiyingying
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-transformer-dev!2280

+++ attention/mla_prolog/op_host/mla_prolog_tiling_check.cpp
@@ -117,12 +117,12 @@ ge::graphStatus MlaPrologTilingCheck::CheckDims() const
     auto socShortName = ascendcPlatform.GetSocVersion();
     if (socShortName == platform_ascendc::SocVersion::ASCEND910_95) {
         OP_CHECK_IF(context_.tokenX.shape->GetStorageShape().GetDimNum() != MLA_PROLOG_DIM_NUM_3,
-            OP_LOGE(context_.opName, "tokenX shape dim num allows only %u, got %u.",
+            OP_LOGE(context_.opName, "tokenX shape dim num allows only %u, got %zu.",
                 MLA_PROLOG_DIM_NUM_3, context_.tokenX.shape->GetStorageShape().GetDimNum()),
             return ge::GRAPH_FAILED);
         OP_CHECK_IF(scenarioInfo_.quantMode_ != QUANT_MODE::NO_QUANT,
-            OP_LOGE(context_.opName, "QUANT_MODE allows only %u, got %u.",
-                QUANT_MODE::NO_QUANT, scenarioInfo_.quantMode_),
+            OP_LOGE(context_.opName, "QUANT_MODE allows only %d, got %d.",
+                static_cast<int>(QUANT_MODE::NO_QUANT), static_cast<int>(scenarioInfo_.quantMode_)),
             return ge::GRAPH_FAILED);
     }
     OP_CHECK_IF(baseShapeInfo_.bSize > MAX_B_SIZE,
@@ -185,7 +185,7 @@ ge::graphStatus MlaPrologTilingCheck::CheckDims() const
         }
         if (*(context_.quantScaleRepoMode) == static_cast<int>(QUANT_SCALE_REPO_MODE::COMBINE)) {
             OP_CHECK_IF(*(context_.tileSize) <= 0,
-                OP_LOGE(context_.opName, "tileSize must > 0, got %u.",
+                OP_LOGE(context_.opName, "tileSize must > 0, got %d.",
                     *(context_.tileSize)),
                 return ge::GRAPH_FAILED);
             supportedDtileSize += baseShapeInfo_.hckvSize / static_cast<uint32_t>(*(context_.tileSize)) * (DTYPE_TO_SIZE.at(ge::DT_FLOAT) / DTYPE_TO_SIZE.at(ge::DT_INT8));
